Use D3 to render tree graph
