+++ package.json
@@ -31,9 +31,11 @@
     "@fortawesome/free-regular-svg-icons": "^6.1.1",
     "@fortawesome/free-solid-svg-icons": "^6.1.1",
     "@fortawesome/vue-fontawesome": "^3.0.0-5",
+    "@types/d3": "^7.4.0",
     "@types/lodash": "^4.14.180",
     "bootstrap": "^4.6.1",
     "clarinet": "^0.12.5",
+    "d3": "^7.4.4",
     "highlight.js": "^11.5.0",
     "humanize-duration": "^3.27.1",
     "lodash": "^4.17.21",
