+++ src/assets/scss/_plan-node.scss
@@ -6,23 +6,12 @@ $bg-color: #fff;
   text-decoration: none;
   color: $text-color;
   display: inline-block;
-  transition: all 0.1s;
   position: relative;
   font-size: $font-size-sm;
   margin-bottom: 4px;
   overflow-wrap: break-word;
   word-wrap: break-word;
-
-  width: 400px;
-  .twod & {
-    width: 240px;
-  }
-  .twod.compact & {
-    width: 140px;
-  }
-  .twod.dot & {
-    width: $compact-width;
-  }
+  width: 95%;
 
   .plan-node-body {
     position: relative;
@@ -153,20 +142,6 @@ $bg-color: #fff;
     }
   }
 
-  .collapse-handle {
-    cursor: pointer;
-    position: absolute;
-    z-index: 1;
-    bottom: 0;
-    right: 0;
-    display: none;
-    background-color: rgba(255, 255, 255, 0.7);
-  }
-
-  &:hover .collapse-handle {
-    display: block;
-  }
-
   &.never-executed .plan-node-body {
     @extend %bg-hatched;
   }
@@ -236,32 +211,6 @@ $bg-color: #fff;
   display: block;
 }
 
-.detailed {
-  width: 400px !important;
-}
-
-ul.node-children.collapsed {
-  .plan-node,
-  .plan-node-body {
-    width: $compact-width !important;
-  }
-  .plan-node {
-    display: block;
-    .plan-node-body {
-      height: math.div($compact-width, 3);
-      * {
-        display: none;
-      }
-    }
-    &:hover .collapse-handle {
-      display: none;
-    }
-  }
-
-  .classic .plan & li::before {
-    top: 7px;
-  }
-  .classic .plan & li:last-child::after {
-    height: 14px;
-  }
+.plan-node-detail {
+  width: 400px;
 }
