+++ src/interfaces.ts
@@ -4,7 +4,6 @@ import type {
   Orientation,
   SortGroupsProp,
   SortSpaceMemoryProp,
-  ViewMode,
 } from "@/enums"
 
 export interface IPlan {
@@ -67,7 +66,8 @@ import { EstimateDirection, NodeProp } from "@/enums"
 
 // Class to create nodes when parsing text
 export class Node {
-  nodeId!: number;
+  nodeId!: number
+  size!: [number, number];
   ["Options"]?: Options;
   ["Timing"]?: Timing;
   ["Settings"]?: Settings;
@@ -116,6 +116,7 @@ export class Node {
     | string
     | string[]
     | undefined
+    | [number, number]
   constructor(type?: string) {
     if (!type) {
       return
@@ -236,7 +237,6 @@ export type ViewOptions = {
   showHighlightBar: boolean
   showPlanStats: boolean
   highlightType: HighlightType
-  viewMode: ViewMode
   orientation: Orientation
   showDiagram: boolean
   diagramWidth: number
