+++ tsconfig.json
@@ -6,7 +6,7 @@
     "paths": {
       "@/*": ["./src/*"]
     },
-    "types": ["node", "jest"]
+    "types": ["node", "jest", "d3"]
   },
 
   "references": [
